billing: refactoring API
