+++ internal/services/billing/di/di.go
@@ -19,10 +19,10 @@ import (
 	payment_application "github.com/batazor/shortlink/internal/services/billing/application/payment"
 	tariff_application "github.com/batazor/shortlink/internal/services/billing/application/tariff"
 	api "github.com/batazor/shortlink/internal/services/billing/infrastructure/api/http"
-	"github.com/batazor/shortlink/internal/services/billing/infrastructure/rpc/balance/v1"
-	"github.com/batazor/shortlink/internal/services/billing/infrastructure/rpc/order/v1"
-	"github.com/batazor/shortlink/internal/services/billing/infrastructure/rpc/payment/v1"
-	"github.com/batazor/shortlink/internal/services/billing/infrastructure/rpc/tariff/v1"
+	"github.com/batazor/shortlink/internal/services/billing/infrastructure/api/rpc/balance/v1"
+	"github.com/batazor/shortlink/internal/services/billing/infrastructure/api/rpc/order/v1"
+	"github.com/batazor/shortlink/internal/services/billing/infrastructure/api/rpc/payment/v1"
+	"github.com/batazor/shortlink/internal/services/billing/infrastructure/api/rpc/tariff/v1"
 	"github.com/batazor/shortlink/internal/services/billing/infrastructure/store"
 	"github.com/batazor/shortlink/pkg/rpc"
 )
