+++ internal/services/billing/di/wire_gen.go
@@ -16,10 +16,10 @@ import (
 	"github.com/batazor/shortlink/internal/services/billing/application/payment"
 	"github.com/batazor/shortlink/internal/services/billing/application/tariff"
 	"github.com/batazor/shortlink/internal/services/billing/infrastructure/api/http"
-	"github.com/batazor/shortlink/internal/services/billing/infrastructure/rpc/balance/v1"
-	"github.com/batazor/shortlink/internal/services/billing/infrastructure/rpc/order/v1"
-	"github.com/batazor/shortlink/internal/services/billing/infrastructure/rpc/payment/v1"
-	"github.com/batazor/shortlink/internal/services/billing/infrastructure/rpc/tariff/v1"
+	"github.com/batazor/shortlink/internal/services/billing/infrastructure/api/rpc/balance/v1"
+	"github.com/batazor/shortlink/internal/services/billing/infrastructure/api/rpc/order/v1"
+	"github.com/batazor/shortlink/internal/services/billing/infrastructure/api/rpc/payment/v1"
+	"github.com/batazor/shortlink/internal/services/billing/infrastructure/api/rpc/tariff/v1"
 	"github.com/batazor/shortlink/internal/services/billing/infrastructure/store"
 	"github.com/batazor/shortlink/pkg/rpc"
 	"github.com/google/wire"
