+++ internal/services/billing/infrastructure/api/rpc/balance/v1/balance_rpc.proto
@@ -1,6 +1,6 @@
 syntax = "proto3";
 
-package infrastructure.rpc.balance.v1;
+package infrastructure.api.rpc.balance.v1;
 
 option go_package = "github.com/batazor/shortlink/internal/billing/infrastructure/balance_rpc";
