+++ internal/services/billing/infrastructure/api/rpc/order/v1/order_rpc.proto
@@ -1,6 +1,6 @@
 syntax = "proto3";
 
-package infrastructure.rpc.order.v1;
+package infrastructure.api.rpc.order.v1;
 
 option go_package = "github.com/batazor/shortlink/internal/billing/infrastructure/order_rpc";
