+++ internal/services/billing/infrastructure/api/rpc/payment/v1/billing_rpc.proto
@@ -1,6 +1,6 @@
 syntax = "proto3";
 
-package infrastructure.rpc.payment.v1;
+package infrastructure.api.rpc.payment.v1;
 
 option go_package = "github.com/batazor/shortlink/internal/billing/infrastructure/payment_rpc";
