+++ internal/services/billing/infrastructure/api/rpc/tariff/v1/tariff_rpc.proto
@@ -1,6 +1,6 @@
 syntax = "proto3";
 
-package infrastructure.rpc.tariff.v1;
+package infrastructure.api.rpc.tariff.v1;
 
 option go_package = "github.com/batazor/shortlink/internal/billing/infrastructure/tariff_rpc";
